pyproject.toml
[project]
name = "legion-aaas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "ollama>=0.6.1",
    "pygithub>=2.9.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "rich>=14.3.3",
]
